requirements.txt
asgiref==3.9.1
attrs==25.3.0
Brotli==1.1.0
click==8.2.1
dj-database-url==3.0.1
Django==5.2.5
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
uvicorn==0.35.0
whitenoise==6.9.0
